Delete comments in SCSS files, delete unused LOS service, disable test files generation in schematics field, improve cellRenderer functions and move them to utils.ts, implement new number rounding librabry
