+++ src/app/length-of-stay/length-of-stay.component.ts
@@ -1,7 +1,8 @@
 import { Component, inject, computed } from '@angular/core';
-import { AgGridAngular } from 'ag-grid-angular'; // Angular Data Grid Component
-import { LOSTableColDefs, numberRoundingFormatter } from './col-defs-los';
-import { ColDef, ColGroupDef, CellValueChangedEvent, GridApi, GridReadyEvent } from 'ag-grid-community'; // Column Definition Type Interface
+import { AgGridAngular } from 'ag-grid-angular';
+import { LOSTableColDefs } from './col-defs-los';
+import { numberRoundingFormatter } from '../utils';
+import { ColDef, ColGroupDef, CellValueChangedEvent, GridApi, GridReadyEvent } from 'ag-grid-community';
 import 'ag-grid-community/styles/ag-grid.css';
 import 'ag-grid-community/styles/ag-theme-quartz.css';
 import { HourlyDataService } from '../hourly-data.service';
@@ -33,7 +34,7 @@ export class LengthOfStayComponent {
     // autoHeight: true,
     minWidth: 100,
     flex: 1,
-    cellRenderer: (params: any) => numberRoundingFormatter(params),
+    cellRenderer: numberRoundingFormatter,
   };
 
   readonly columnDefs: (ColDef | ColGroupDef)[] = LOSTableColDefs;
