--- src/app/los-data.service.ts
@@ -1,468 +1,0 @@
-import { Injectable, signal } from '@angular/core';
-import { CONSTANTS } from './constants';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class LOSDataService {
-  readonly rowData = [
-    {
-      godzina: '0-1',
-      oczekiwaneWizyty: 3.23,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 3.23,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 3.23,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.3,
-      zmiennoscPielegniarkaWq: 0.09,
-      zmiennoscLekarzLq: 0.01,
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.1,
-    },
-    {
-      godzina: '1-2',
-      oczekiwaneWizyty: 2.39,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 2.39,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 2.39,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.07,
-      zmiennoscPielegniarkaWq: 0.03,
-      zmiennoscLekarzLq: 0.01,
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.03,
-    },
-    {
-      godzina: '2-3',
-      oczekiwaneWizyty: 2.14,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 2.14,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 2.14,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.04,
-      zmiennoscPielegniarkaWq: 0.02,
-      zmiennoscLekarzLq: 'df',
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.02,
-    },
-    {
-      godzina: '3-4',
-      oczekiwaneWizyty: 1.83,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 1.83,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 1.83,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.02,
-      zmiennoscPielegniarkaWq: 0.01,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.01,
-    },
-    {
-      godzina: '4-5',
-      oczekiwaneWizyty: 1.7,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 1.7,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 1.7,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.01,
-      zmiennoscPielegniarkaWq: 0.01,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.01,
-    },
-    {
-      godzina: '5-6',
-      oczekiwaneWizyty: 1.67,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 1.67,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 1.67,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.01,
-      zmiennoscPielegniarkaWq: 0.01,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.01,
-    },
-    {
-      godzina: '6-7',
-      oczekiwaneWizyty: 1.8,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 1.8,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 1.8,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.01,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.01,
-    },
-    {
-      godzina: '7-8',
-      oczekiwaneWizyty: 2.96,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 2.96,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 2.96,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.01,
-      opoznienie: 0.01,
-    },
-    {
-      godzina: '8-9',
-      oczekiwaneWizyty: 5.66,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 5.66,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 5.66,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.19,
-      zmiennoscPielegniarkaWq: 0.03,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.02,
-      opoznienie: 0.02,
-    },
-    {
-      godzina: '9-10',
-      oczekiwaneWizyty: 9.25,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 9.25,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 9.25,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 1.47,
-      zmiennoscPielegniarkaWq: 0.34,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.16,
-      opoznienie: 0.3,
-    },
-    {
-      godzina: '10-11',
-      oczekiwaneWizyty: 12.15,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 8,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 12.15,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 12.15,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.0,
-    },
-    {
-      godzina: '11-12',
-      oczekiwaneWizyty: 13.25,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 10.99,
-      opoznieniePielegniarkaObsluga: 13.25,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 13.25,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.67,
-    },
-    {
-      godzina: '12-13',
-      oczekiwaneWizyty: 12.25,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 12.25,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 12.25,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.38,
-    },
-    {
-      godzina: '13-14',
-      oczekiwaneWizyty: 10.52,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 10.52,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 10.52,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.57,
-    },
-    {
-      godzina: '14-15',
-      oczekiwaneWizyty: 9.44,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 9.44,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 9.44,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.38,
-    },
-    {
-      godzina: '15-16',
-      oczekiwaneWizyty: 9.73,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 9.73,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 9.73,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.58,
-    },
-    {
-      godzina: '16-17',
-      oczekiwaneWizyty: 10.29,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 10.29,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 10.29,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.67,
-    },
-    {
-      godzina: '17-18',
-      oczekiwaneWizyty: 10.1,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 10.1,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 10.1,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.38,
-    },
-    {
-      godzina: '18-19',
-      oczekiwaneWizyty: 9.34,
-      liczbaPielegniarek: 7,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 9.34,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 9.34,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.21,
-    },
-    {
-      godzina: '19-20',
-      oczekiwaneWizyty: 8.57,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 8.84,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 8.57,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 8.57,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.11,
-    },
-    {
-      godzina: '20-21',
-      oczekiwaneWizyty: 7.61,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 7.61,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 7.61,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.1,
-    },
-    {
-      godzina: '21-22',
-      oczekiwaneWizyty: 6.33,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 6.33,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 6.33,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.09,
-    },
-    {
-      godzina: '22-23',
-      oczekiwaneWizyty: 5.08,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 5.08,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 5.08,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.04,
-    },
-    {
-      godzina: '23-24',
-      oczekiwaneWizyty: 4.08,
-      liczbaPielegniarek: 4,
-      wydajnoscPielegniarek: 5.05,
-      liczbaLekarzy: 6,
-      wydajnoscLekarzy: 8.24,
-      opoznieniePielegniarkaObsluga: 4.08,
-      opoznieniePielegniarkaKolejka: 0.0,
-      opoznieniePielegniarkaOczekiwanie: 0.0,
-      opoznienieLekarzObsluga: 4.08,
-      opoznienieLekarzKolejka: 0.0,
-      opoznienieLekarzOczekiwanie: 0.0,
-      zmiennoscPielegniarkaLq: 0.0,
-      zmiennoscPielegniarkaWq: 0.0,
-      zmiennoscLekarzLq: 0.0,
-      zmiennoscLekarzWq: 0.0,
-      opoznienie: 1.0,
-    },
-  ];
-
-  constructor() {}
-}
