+++ src/app/static-table/col-defs-static.ts
@@ -1,22 +1,4 @@
 import { ColDef, ColGroupDef } from 'ag-grid-community';
-import { CellClassParams } from 'ag-grid-community';
-
-const DECIMAL_PLACES = 100; // 10 for 1, 100 for 2, 1000 for 3 ...
-export function numberRoundingFormatter(params: any) {
-  if (params.data['id'] === 11 && params.value === 777) {
-    return '***';
-  }
-
-  if (params.value === null) {
-    return '';
-  }
-
-  if (!isNaN(Number(params.value))) {
-    return `${Math.round(params.value * DECIMAL_PLACES) / DECIMAL_PLACES}`;
-  }
-
-  return params.value;
-}
 
 export const staticTableColDefs: (ColDef | ColGroupDef)[] = [
   {
